Move hyperparams optim in dedicated module
